cloud-detection-code/requirements.txt
absl-py==1.0.0
alabaster==0.7.12
algopy==0.5.7
antialiased-cnns==0.3
anybadge==1.9.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.8.6
astropy==5.0.4
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
autograd==1.4
Babel==2.10.1
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bleach==5.0.0
build==0.8.0
cachetools==5.1.0
Cartopy==0.20.2
certifi==2021.10.8
cffi==1.15.0
cftime==1.6.0
charset-normalizer==2.0.12
commonmark==0.9.1
coverage==6.4.2
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.0
deconfuser==1.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.5
docopt==0.6.1
docopts==0.6.1
docutils==0.17.1
e2cnn==0.2.1
entrypoints==0.4
escnn==0.1.1
execnet==1.9.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
flatbuffers==1.12
focal-loss-pytorch==0.0.2
fonttools==4.33.3
future==0.18.2
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
greenlet==1.1.2
grpcio==1.46.3
h5py==3.6.0
idna==3.3
imageio==2.19.2
imagesize==1.3.0
importlib-metadata==4.11.4
iniconfig==1.1.1
ipykernel==6.14.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jplephem==2.17
jsonschema==4.6.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==23.6.0
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
libclang==14.0.1
lie-learn==0.0.1.post1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mlxtend==0.19.0
mpmath==1.2.1
msgpack==1.0.3
msgpack-numpy==0.4.7.1
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
neovim==0.3.1
nest-asyncio==1.5.5
netCDF4 @ file:///Users/alexmeredith/gpsro-constellation/colocation-finding/netcdf4-python
networkx==2.8.1
notebook==6.4.12
numdifftools==0.9.40
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
patsy==0.5.2
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pipenv==2022.7.24
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycparser==2.21
pyerfa==2.0.0.1
Pygments==2.12.0
pylint==2.11.1
pymanopt==1.0.0
PyNaCl==1.5.0
pynvim==0.4.3
pyparsing==3.0.9
pyproj==3.3.1
pyrsistent==0.18.1
pysftp==0.2.9
pyshp==2.3.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.0.0
qtconsole==5.3.1
QtPy==2.1.0
readme-renderer==35.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.4.4
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
segmentation-models-pytorch==0.3.0
Send2Trash==1.8.0
sgp4==2.21
Shapely==1.8.2
six==1.15.0
skyfield==1.42
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
statsmodels==0.13.2
sympy==1.10.1
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.0
tensorflow-metadata==1.8.0
tensorflow-metal==0.5.0
tensorpack==0.11
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
torch==1.11.0
torchgeometry==0.1.2
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
twine==4.0.1
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.16.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.13.3
zipp==3.8.0
